+++ puremagic/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from __main__ import *
+
+__author__ = "Chris Griffith"
+__version__ = "1.2"
